Add files via upload
the word counter project is a Python program that helps you count the number of words in a given text. It's a handy tool for analyzing text documents, essays, or even social media posts. The program takes input from the user, processes the text, and then displays the total word count. It's a simple yet useful project that can be a great learning experience for Python beginners like me . Feel free to check it out on my GitHub repository!"